+++ tests/testthat/get_last_news_error/geoservices.ign.fr/actualites/rss.xml.R
@@ -1,5 +1,0 @@
-structure(list(method = "GET", url = "https://geoservices.ign.fr/actualites/rss.xml", 
-    status_code = 503L, headers = structure(list(`cache-control` = "no-store", 
-        `content-length` = "191", `content-type` = "application/json", 
-        connection = "close", Date = "Sun, 19 Mar 2023 11:23:15 UTC"), class = "httr2_headers"), 
-    body = charToRaw("{ \"errors\" : [ { \"status\" : \"503\", \"title\" : \"Service unavailable\", \"detail\" : \"No server is available to handle this request. Unique ID: D8A35C8D7FAADC81C086880B01BB6416F0A019A754122F\" } ] }")), class = "httr2_response")