+++ tests/testthat/test-get_layers_metadata.R
@@ -15,8 +15,24 @@ with_mock_dir("get_layers_metada_wms", {
 
       res <- get_layers_metadata("ortho", "wms")
 
-      expect_equal(dim(res), c(11, 3))
+      expect_equal(dim(res), c(14, 3))
       expect_equal(names(res),
                    c("Name", "Title", "Abstract"))
       })
+})
+
+with_mock_dir("get_layers_metada_null", {
+   test_that("get_layers_metada_null", {
+
+      expect_warning(get_layers_metadata("adresse", "wms"),
+                     "There's no wms resources")
+   })
+})
+
+test_that("get_layers_metada_error", {
+
+   expect_error(get_layers_metadata("bad_apikey", "wms"), "`apikey` must be one of :")
+   expect_error(get_layers_metadata(NA, "wms"), "`apikey` must be of class")
+   expect_error(get_layers_metadata(NULL, "wms"), "`apikey` must be of class")
+   expect_error(get_layers_metadata(c("altimetrie", "adresse"), "wms"), "Only one")
 })